your commit message

+++ components/BlogList.vue
@@ -18,26 +18,34 @@
                 <tr v-for="(blog,index) of blogs" v-bind:key="blog.id" v-bind:title="blog.title">
                     <td>{{blog.id}}</td>
                     <td>{{blog.title}}</td>
-                    <td v-if="blog.category == 0">Thời sự</td>
-                    <td v-else-if="blog.category == 1">Thể thao</td>
-                    <td v-else-if="blog.category == 2">Văn hóa</td>
-                    <td v-else-if="blog.category == 3">Nghệ thuật</td>
-                    <td v-else>Chính trị</td>
-                    <td v-if="blog.public === false">Không hiển thị</td>
-                    <td v-else-if="blog.public === true">Hiển thị</td>
+                    <td>
+                        <div
+                        v-for="(cate, index) in dataCate"
+                        :key="index"
+                        >
+                        <p v-if="index == blog.category">{{cate}}</p>
+
+                        </div>
+                    </td>
+                    <td v-if="blog.public == true ">Yes</td>
+                    <td v-if="blog.public == false ">No</td>
                     <td>
                         <ul>
-                            <div v-for="position of blog.position" >
-                                <li v-if="position == 1"> Hà Nội</li>
-                                <li v-if="position == 2"> Châu Âu</li>
-                                <li v-if="position == 3"> Châu Á</li>
-                                <li v-if="position == 4"> Châu Mỹ</li>
-                            </div>
+                        <div
+                            v-for="(posit) of blog.position"
+                            v-bind:key="posit"
+                        >
+                            <li
+                            v-for="(pos, index) in dataPos"
+                            :key="index"
+                            >
+                            <p v-if="index + 1 == posit">{{pos}}</p>
+                            </li>
+                        </div>
                         </ul>
-                        
                     </td>
                     <td>{{blog.data_pubblic}}</td>
-                    <td><button type="button" class="btn btn-primary"><a v-bind:href="'/edit/'+ blog.id" title="">Edit</a></button></td>
+                    <td><button type="button" class="btn btn-primary"><a v-bind:href="'/'+ blog.id" title="">Edit</a></button></td>
                     <td><button type="button" class="btn btn-danger" v-on:click="fireDelete(blog.id, index)">Delete</button></td>
                 </tr>
             </tbody>
@@ -46,16 +54,33 @@
 </template>
 
 <script>
-
+import { DATA_CATE } from '@/store/const.js'
+import { DATA_POS } from '@/store/const.js'
 import axios from "axios";
 export default {
     props:['blogs'],
+    data() {
+        return {
+        DATA_CATE: DATA_CATE,
+        DATA_POS: DATA_POS,
+        }
+    },
+    computed: {
+        dataCate: function () {
+        return this.DATA_CATE
+        },
+
+        dataPos: function () {
+        return this.DATA_POS
+        }
+    },
     methods: {       
         fireDelete(id, index) {      
-            console.log(index),
+            if(confirm("Bạn có chắc chắn muốn xóa không ?")){
             axios.delete('http://localhost:4000/blogs/'+id).then((res) => {
                 this.blogs.splice(index, 1);
-            })    
+            }) 
+            }   
         },    
     }
 }
